requirements.txt
# Core
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.8

# Torch (CPU)
torch==2.2.2

# Kafka & S3
kafka-python>=2.0.2
boto3>=1.34
botocore>=1.34

# ML explain 
shap>=0.44; python_version >= "3.10" 

# System utils
psutil>=5.9

# Kubeflow Pipelines client
kfp>=2.6

# Others
pyyaml>=6.0
